Moves BackedUpDirectory to common
