+++ kifi-backend/modules/common/app/com/keepit/common/package.scala
@@ -5,6 +5,7 @@ import java.io._
 import org.apache.commons.io.{IOUtils, FileUtils}
 import scala.collection.mutable.{SynchronizedSet, HashSet, Set => MutableSet, ListBuffer}
 import java.util.zip.{GZIPInputStream, GZIPOutputStream}
+import java.util.concurrent.atomic.AtomicBoolean
 
 package object common {
 
@@ -114,6 +115,38 @@ package object common {
       val inputStream = FileUtils.openInputStream(tarGz)
       try { uncompress(inputStream, destDir) }
       finally { inputStream.close() }
+    }
+  }
+
+  trait BackedUpDirectory {
+    def getDirectory(): File
+    def scheduleBackup(): Unit
+    def cancelBackup(): Unit
+    def doBackup(): Boolean
+    def restoreFromBackup(): Unit
+  }
+
+  trait ArchivedDirectory extends BackedUpDirectory {
+    protected def getArchive(): File
+    protected def saveArchive(archive: File): Unit
+
+    private val shouldBackup = new AtomicBoolean(false)
+    def scheduleBackup() = shouldBackup.set(true)
+    def cancelBackup() = shouldBackup.set(false)
+    def doBackup() = if (shouldBackup.getAndSet(false)) {
+      val dir = getDirectory()
+      val tarGz = IO.compress(dir)
+      saveArchive(tarGz)
+      tarGz.delete()
+      true
+    } else false
+
+    def restoreFromBackup(): Unit = {
+      val dir = getDirectory()
+      val tarGz = getArchive()
+      FileUtils.deleteDirectory(dir)
+      IO.uncompress(tarGz, dir.getParentFile.getAbsolutePath)
+      tarGz.delete()
     }
   }
