+++ kifi-backend/modules/search/app/com/keepit/search/index/IndexDirectory.scala
@@ -2,46 +2,12 @@ package com.keepit.search.index
 
 import java.io.File
 import com.keepit.common.store._
-import com.keepit.common.IO
+import com.keepit.common.{ArchivedDirectory, BackedUpDirectory}
 import org.apache.lucene.store.{RAMDirectory, MMapDirectory, Directory}
 import com.amazonaws.services.s3.AmazonS3
 import com.keepit.common.store.S3Bucket
 import com.keepit.common.logging.Logging
 import com.keepit.common.logging.AccessLog
-import java.util.concurrent.atomic.AtomicBoolean
-import org.apache.commons.io.FileUtils
-
-trait BackedUpDirectory {
-  def getDirectory(): File
-  def scheduleBackup(): Unit
-  def cancelBackup(): Unit
-  def doBackup(): Boolean
-  def restoreFromBackup(): Unit
-}
-
-trait ArchivedDirectory extends BackedUpDirectory {
-  protected def getArchive(): File
-  protected def saveArchive(archive: File): Unit
-
-  private val shouldBackup = new AtomicBoolean(false)
-  def scheduleBackup() = shouldBackup.set(true)
-  def cancelBackup() = shouldBackup.set(false)
-  def doBackup() = if (shouldBackup.getAndSet(false)) {
-    val dir = getDirectory()
-    val tarGz = IO.compress(dir)
-    saveArchive(tarGz)
-    tarGz.delete()
-    true
-  } else false
-
-  def restoreFromBackup(): Unit = {
-    val dir = getDirectory()
-    val tarGz = getArchive()
-    FileUtils.deleteDirectory(dir)
-    IO.uncompress(tarGz, dir.getParentFile.getAbsolutePath)
-    tarGz.delete()
-  }
-}
 
 trait IndexStore extends ObjectStore[IndexDirectory, File]
